requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_pytorch_select=0.2=gpu_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
ca-certificates=2022.9.24=ha878542_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py37h74dc2b5_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.9=h6a678d5_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
future=0.18.2=py37_1
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py37h295c915_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.2=py37h06a4308_0
matplotlib-base=3.5.2=py37hf590b9c_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pillow=6.2.1=py37h34e0f95_0
pip=22.2.2=py37h06a4308_0
ply=3.11=py37_0
pycparser=2.21=pyhd3eb1b0_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.3.1=cuda100py37h53c1284_0
pytorch-gpu=1.3.1=0
pytorch-model-summary=0.1.1=py_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.1.2=h7f8727e_1
setuptools=63.4.1=py37h06a4308_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h5082296_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torch-summary=1.4.5=pypi_0
torchvision=0.4.2=cuda100py37hecfc37a_0
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=py37h06a4308_0
typing-extensions=4.3.0=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
tzdata=2022c=h04d1e81_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
zlib=1.2.12=h5eee18b_3
zstd=1.5.2=ha4553b6_0
